refactor: replace 'dialoguer' with 'forge_select' for user interaction prompts
